feat: add `endpoint-readit-add-summary-to-eval-queue` CLI and refactor GitHub GraphQL logic
Implemented the evaluation queue registration CLI and refactored shared
GitHub Project V2 GraphQL mutations into a common module.

Changes:
- Created `src/endpoint/readit/github.py` with reusable `AddProjectV2DraftIssue` and `UpdateTextFieldValue` classes.
- Implemented `src/endpoint/readit/app/add_summary_to_eval_queue.py` using shared logic.
- Updated `src/endpoint/readit/app/send_to_queue_v2.py` to use shared logic.
- Registered the new CLI in `pyproject.toml`.
- Added `src/endpoint/readit/__init__.py` for package support.
