+++ src/endpoint/readit/app/send_to_queue_v2.py
@@ -1,73 +1,20 @@
 import click
-
 from gql import Client
-from gql import gql
-from gql.transport.requests import RequestsHTTPTransport as HTTPTransport
 
 import json
 from logging import getLogger
 import os
-from typing import NewType
 
 from endpoint.readit.core import Page
+from endpoint.readit.github import ProjectItemID
+from endpoint.readit.github import AddProjectV2DraftIssue
+from endpoint.readit.github import UpdateTextFieldValue
+
+from gql.transport.requests import RequestsHTTPTransport as HTTPTransport
 
 
 logger = getLogger(__name__)
 logger.setLevel(os.environ.get("ENTRYPOINT_LOG_LEVEL", "INFO").upper())
-
-ProjectItemID = NewType("ProjectItemID", str)
-
-
-class AddProjectV2DraftIssue:
-    QUERY = gql("""
-    mutation ($projectId: ID!, $title: String!, $body: String!) {
-      op: addProjectV2DraftIssue(input: {
-        projectId: $projectId,
-        title: $title,
-        body: $body,
-      }) { item: projectItem { id } }
-    }
-    """)
-
-    def __init__(self, *, projectId: str, title: str, body: str):
-        self._values = {
-            "projectId": projectId,
-            "title": title,
-            "body": body,
-        }
-
-    def execute(self, client) -> ProjectItemID:
-        result = client.execute(self.QUERY, variable_values=self._values)
-        logger.debug(result)
-        return result["op"]["item"]["id"]
-
-
-class UpdateTextFieldValue:
-    QUERY = gql("""
-    mutation ($projectId: ID!, $itemId: ID!, $fieldId: ID!, $value: String!) {
-      updateProjectV2ItemFieldValue(input: {
-        projectId: $projectId,
-        itemId: $itemId,
-        fieldId: $fieldId,
-        value: { text: $value }
-      }) { item: projectV2Item { id } }
-    }
-    """)
-
-    def __init__(
-        self, *, projectId: str, itemId: ProjectItemID, fieldId: str, value: str
-    ):
-        self._values = {
-            "projectId": projectId,
-            "itemId": str(itemId),
-            "fieldId": fieldId,
-            "value": value,
-        }
-
-    def execute(self, client) -> ProjectItemID:
-        result = client.execute(self.QUERY, variable_values=self._values)
-        logger.debug(result)
-        pass
 
 
 class Queue:
